Support reading from a stream, such as a .zip file
A new base class StreamDBF has been added that takes a file-like object
rather than a file name. StreamDBF only requires the 'read' method on
the provided object. Some limitations arise from this - for example it
is not possible to read the records more than once. However for some
purposes, such as extracting data from a .dbf file to upload elsewhere,
it is sufficient.

The DBF class is now a descendant of StreamDBF.
